+++ addons/guide/editor/utils.gd
@@ -20,3 +20,15 @@ static func is_inline(resource:Resource) -> bool:
 	if resource == null:
 		return false
 	return resource.resource_path.contains("::") or resource.resource_path == ""
+	
+	
+## Checks if the given node is somewhere in the currently edited scene.	
+static func is_node_in_edited_scene(node:Node) -> bool:
+	if not is_instance_valid(node):
+		return false
+	
+	var scene_root := EditorInterface.get_edited_scene_root()
+	if not is_instance_valid(scene_root):
+		return false
+		
+	return (node == scene_root) or scene_root.is_ancestor_of(node)